feat: Tags goodies for issue #4
Issue #4's objective is to support adding default tags to elements and
relationships. For example, a `Person` should have 'Element' and 'Person' as
its default tags. Similarly, a `_Relationship` should have 'Relationship' tag
by default.

On top of this, now the interfaces takes a `Set[str]` as tags instead of a
`List[str]` for obvious reason.

Elements and Relationships should now also has the `tags` property to return
the set of tags.