Add support for implied relationships
This commit allows implied relationships to be created when it is enabled.

Also, found a bug where source id is `None` when using the `uses` method.
This is because the source id is only defined in the initialization of `_UsesFrom`, which the syntax of the form `a >> R >> b` uses, but not in the `uses` method. Hence the refactoring of the `uses` method to use `_UsesFrom`.