feat(dsl): add type checks for relationships
This also adds another test file specificaly for type hints.

This is especially useful for type hinting the possible destination element for
specific source element in a relationship. For example, the `Person` element,
can only have either a `Person`, `SoftwareSystem`, `Container`, or `Component`
as a destination; not `DeploymentNode` or `InfrastructureNode`.

Likewise, the `DeploymentNode` element can only have a `DeploymentNode` as its
destination in a relationship.

This commit is added for #67 in anticipation that typehints may be
false-positive to the relationship between `Person` and `DeploymentNode`, for
example.

chore(ci): update build.yml

- Tests didn't trigger when src changed. This commit tries to fix that.