feat(dsl): add get_workspace method to FluentSink
This method allows us to get the workspace from a FluentSink instance, which was
what many of the tests prevoiusly assumed were possible. The test failure wasn't
caught in the CI because of bad test trigger check (fixed in 45e14465).

As such, this method may be deprecated in the future, but for now it is
necessary to be kept.