feat(dsl): add type checks for relationships (#68)
* feat(dsl): add type checks for relationships

This also adds another test file specificaly for type hints.

This is especially useful for type hinting the possible destination element for
specific source element in a relationship. For example, the `Person` element,
can only have either a `Person`, `SoftwareSystem`, `Container`, or `Component`
as a destination; not `DeploymentNode` or `InfrastructureNode`.

Likewise, the `DeploymentNode` element can only have a `DeploymentNode` as its
destination in a relationship.

This commit is added for #67 in anticipation that typehints may be
false-positive to the relationship between `Person` and `DeploymentNode`, for
example.

chore(ci): update build.yml

- Tests didn't trigger when src changed. This commit tries to fix that.

* ci: [skip-ci] Update version to 0.0.7.dev0

* feat(dsl): add get_workspace method to FluentSink

This method allows us to get the workspace from a FluentSink instance, which was
what many of the tests prevoiusly assumed were possible. The test failure wasn't
caught in the CI because of bad test trigger check (fixed in 45e14465).

As such, this method may be deprecated in the future, but for now it is
necessary to be kept.

* fix(dsl): type hints when rhs in a list

* ci: [skip-ci] Update version to 0.0.7.dev3

* chore(ci): use the correct ref of PR source

* ci: [skip-ci] Update version to 0.0.7.dev5

---------

Co-authored-by: GitHub Actions <actions@github.com>