feat(dsl): Ensure all tests in `test_dsl.py` are passed
The main issue now is about the test `test_source_destinations_in_dsl_elements`
where the sources and destinations list of the affected elements are being
duplicated whenever we're creating implied relationships.

For now, we'll just check for duplicate before adding new source/destination to
the elements when creating a new relationship. But a more neat solution would be
to use the `set` data structure to maintain the sources and destinations.