feat(dsl): Improve relationship explorer
The existing `Explorer.walk_relationships` requires the recreation of the
relationships -- resulting in a "fake" relationship. This was good enough for
read only operations, but not sufficient when we need to write or update a
property of existing relationship (e.g., updating its tag for styling purposes).

This commit improves this by also having the `DslElement` to track its
associated relationships (i.e., via the `relationships` property).