Add features to apply styles to elements and relationships in views (#75)
* feat(dsl): Add a feature to apply element styles to views

- Had to refactor `Group` to allow us to get the group's full name or path
  even when it's outside the `Group` context. (Which should be the case anyway!)
- On top of that, we added a new `Color` class to handle the color
  representation of the elements.
- Other than using basic tag strings and group to select the elements to style,
  we can also select the elements by their types, or by using a Callable method
  to get matching elements, or simply by directly passing the elements to style.

* fix(dsl): Fix element styling using callable

When using a callable to set the style of matched elements, all elements in the
workspace are applied the same style. This commit fixes the issue by ensuring
that the style is applied only to the matched elements.

* feat(dsl): Add relationships property to DslElement

This makes it so that we can track back the relationships
created with the associated DslElement.

Might also be useful to replace the "fake" relationships used in the
`Explorer.relationships` method.

* feat(dsl): Improve relationship explorer

The existing `Explorer.walk_relationships` requires the recreation of the
relationships -- resulting in a "fake" relationship. This was good enough for
read only operations, but not sufficient when we need to write or update a
property of existing relationship (e.g., updating its tag for styling purposes).

This commit improves this by also having the `DslElement` to track its
associated relationships (i.e., via the `relationships` property).

* feat(dsl): Add support for styling relationships

Other than the basic Structurizr DSL way of styling relationships with tags,
we can also use the following objects to tell Structurizr which relations to style:
- The relationship objects
- The group within which the relationship is defined
- The predicates (`Callable[[Workspace, Relationship], bool]`) that can be used
  to logically determine exactly which relationships to style
- A `None`, which means all relationships in the workspace

* fix(dsl): Fix duplicating relationships with `walk_relationships`

Now we just add the relationship for the source element, not the target element.

* feat(dsl): Change parameter name for selecting elements on relationships

When selecting elements and/or relationships for styling with `StyleElements`
and `StyleRelationships`, the parameter name has been changed from
`elements`/`relationships` to `on`. This is to make it more general (selected
"elements" and "relationships" are not always exactly elements or relationships
-- they may also be callables or groups.)

* fix(test): Fix tests

Since now `DslElement.relationships` only track the relationships for the source
elements, we need to update the tests to reflect this change.

* ci: [skip-ci] Update version to 0.0.9.dev7

---------

Co-authored-by: GitHub Actions <actions@github.com>