feat(expression): improve expressions (#79)
* feat(expression): improve expressions

So far, the callables for `Expression`s, used for filtering and selecting
elements and relationships, uses the same `Workspace` class we use for creating
the models. This was not the case for `Expression.Element` (for filtering
`DslElement` and `Expression.Relationship` for filtering `DslRelationship`).

This commit make the following changes:
- Rename `Element`, and `Relationship` to `ElementExpression` and
  `RelationshipExpression` respectively to avoid confusion with the
  `DslElement` and `DslRelationship` classes.
- Create a new `Expression` class for `Workspace` called `WorkspaceExpression`
  that uses the `Workspace` class for filtering and selecting elements and
  relationships.

Now, you can't do something like `workspace.apply_view(...)` in a
`Callable[[WorkspaceExpression, ElementExpression], bool]` predicate because
the `apply_view` method is not available in the `WorkspaceExpression` class.

* ci: [skip-ci] Update version to 0.0.13.dev0

---------

Co-authored-by: GitHub Actions <actions@github.com>