feat(dsl): (WIP) Add DeploymentView
The implementation for this view can be quite complex, requiring recursive
traversal of the deployment nodes models with upstream context awareness in
order to include the correct elements.

For example, depending on whether or not we specify a software system to the
view, and whether or not the software system instance is defined, we may or may
not include the container instances in the view. See the DeploymentView test
scripts for more details.

Note: InfrastructureNode hasn't been implemented yet, so this is a work in
progress.