Add support for Deployment Node, Infrastructure Node, Software System Instance, and Container Instance (#80)
* chore: upgrade to python 3.10

* feat(dsl): add support for deployment nodes and container instances

Follows example from the cookbook:
https://docs.structurizr.com/dsl/cookbook/deployment-groups/

* feat(dsl): add support for more deployment models

Add support for the following DSL elements:
- `InfrastructureNode`
- `DeploymentGroup`
- `SoftwareSystemInstance`

* feat(dsl): add tags params to deployment elements

* feat(dsl): add relationships support nodes and instances

This commit allows the definition of relationships between nodes and instances:
- Deployment Node --> Deployment Node
- Infrastructure Node --> Deployment Node, Infrastructure Node,
                          Software System Instance, Container Instance
- Software System Instance --> Infrastructure Node
- Container Instance --> Infrastructure Node

Anything that's outside of the defined relationships above will cause `mypy` to
complain.

There's an unfortunate impact of this implementation to `ElementExpression`
returned by `Expression(...).elements()` filter:
- The `DslElementRelationOverrides` which is used to define the legal relations
  requires the classes to inherit `DslElement`.
- `SoftwareSystemInstance` and `ContainerInstance` does not have a `name`
  property, so not all `DslElement` returned by the filter will have a name property.
  As shown in the changes to `test_expression.py`, we'll have to use
  `isinstance(...)` or `hasattr(...)` to make mypy happy when we use
  `Expression(...).elements()`.

* ci: make sure pytest also tests with mypy

* fix(dsl): fix issues with tags

Tags for deployment nodes and instances aren't registered correctly.

This also prompted me to add a test for ensuring tags for elements in the DSL.

* feat(dsl): ensure Explorer includes deployment nodes and instances

* feat(dsl): (WIP) Add DeploymentView

The implementation for this view can be quite complex, requiring recursive
traversal of the deployment nodes models with upstream context awareness in
order to include the correct elements.

For example, depending on whether or not we specify a software system to the
view, and whether or not the software system instance is defined, we may or may
not include the container instances in the view. See the DeploymentView test
scripts for more details.

Note: InfrastructureNode hasn't been implemented yet, so this is a work in
progress.

* ci: [skip-ci] Update version to 0.0.13

* refactor: refactor tests/test_views.py

* chore: remove old comments

* feat(view): add test for deployment groups

* feat(views): add support for infrastructure nodes

* feat(dsl): add implied instance relationships

The implied instance relationships are used for deployment views.

For example, if we have containers A and B, and we have a relationship A --> B,
and each A and B have instances A1 and A2, and B1 and B2, then we need to create
the relationships:
- A1 --> B1
- A1 --> B2
- A2 --> B1
- A2 --> B2

* chore: remove print statements from dsl.py

* ci: [skip-ci] Update version to 0.0.14.dev14

---------

Co-authored-by: GitHub Actions <actions@github.com>